First upload
This is a bit of a long time coming, but I was fighting with a design concept and I think I'm off to a solid start. There's no broken code, I did lift and shift some of the cr1.0 stuff since it's still applicable, just some moderate re-tooling. Figure now is as good a time as any to commit to the repo so that if god forbid I suffer some kind of catastrophic event with my laptop all isn't lost.