v2.0 main done
Pretty well finished I think! I need to get the contact page in place now, but that's fine, need to work through what that will look like (considering eschewing the navbar for just a form with a return to home button if the user decides not to use the form).